+++ src/Microsoft.DotNet.Darc/tests/Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests/VmrTestsOneTimeSetUp.cs
@@ -21,7 +21,9 @@ public class VmrTestsOneTimeSetUp
 {
     public static readonly LocalPath TestsDirectory;
     public static readonly LocalPath CommonVmrPath;
-    public static readonly LocalPath CommonPrivateRepoPath;
+    public static readonly LocalPath CommonProductRepoPath;
+    public static readonly LocalPath FirstInstallerDependencyPath;
+    public static readonly LocalPath SecondInstallerDependencyPath;
     public static readonly LocalPath CommonDependencyPath;
     public static readonly LocalPath CommonInstallerPath;
     public static readonly LocalPath CommonExternalRepoPath;
@@ -35,10 +37,12 @@ public class VmrTestsOneTimeSetUp
         ResourcesPath = new NativePath(Path.Join(Path.GetDirectoryName(assembly.Location), "Resources"));
         TestsDirectory = new NativePath(Path.GetTempPath()) / "_vmrTests" / Path.GetRandomFileName();
         CommonVmrPath = TestsDirectory / Constants.VmrName;
-        CommonPrivateRepoPath = TestsDirectory / Constants.ProductRepoName;
+        CommonProductRepoPath = TestsDirectory / Constants.ProductRepoName;
         CommonExternalRepoPath = TestsDirectory / Constants.SubmoduleRepoName;
         CommonDependencyPath = TestsDirectory / Constants.DependencyRepoName;
         CommonInstallerPath = TestsDirectory / Constants.InstallerRepoName;
+        FirstInstallerDependencyPath = TestsDirectory / Constants.FirstInstallerDependencyName;
+        SecondInstallerDependencyPath = TestsDirectory / Constants.SecondInstallerDependencyName;
     }
 
     [OneTimeSetUp]
@@ -50,23 +54,33 @@ public class VmrTestsOneTimeSetUp
         Directory.CreateDirectory(TestsDirectory / Constants.VmrName / VmrInfo.SourcesDir);
         await _gitOperations.InitialCommit(TestsDirectory / Constants.VmrName);
 
-        Directory.CreateDirectory(CommonInstallerPath);
-        Directory.CreateDirectory(CommonInstallerPath / "eng");
-        Directory.CreateDirectory(CommonInstallerPath / Constants.PatchesFolderName / Constants.ProductRepoName);
-
-        File.WriteAllText(CommonInstallerPath / VersionFiles.VersionDetailsXml, Constants.EmptyVersionDetails);
-        await _gitOperations.InitialCommit(CommonInstallerPath);
-
-        var dependenciesMap = new Dictionary<string, List<Dependency>>
+        var repoDependencies = new Dictionary<string, List<Dependency>>
         {
             {Constants.ProductRepoName,  new List<Dependency> {new Dependency(Constants.DependencyRepoName, CommonDependencyPath) } }
         };
 
-        await CreateRepositoryRecursive(CommonPrivateRepoPath, Constants.ProductRepoName, dependenciesMap);
-        File.Copy(ResourcesPath / Constants.ProductRepoFileName, CommonPrivateRepoPath / Constants.ProductRepoFileName, true);
-        await _gitOperations.CommitAll(CommonPrivateRepoPath, "change file content");
+        await CreateRepositoryRecursive(CommonProductRepoPath, Constants.ProductRepoName, repoDependencies);
+        File.Copy(ResourcesPath / Constants.ProductRepoFileName, CommonProductRepoPath / Constants.ProductRepoFileName, true);
+        await _gitOperations.CommitAll(CommonProductRepoPath, "change file content");
 
         await CreateRepositoryRecursive(CommonExternalRepoPath, Constants.SubmoduleRepoName);
+
+        var dependenciesMap = new Dictionary<string, List<Dependency>>
+        {
+            {
+                Constants.InstallerRepoName,  
+                new List<Dependency> 
+                {
+                    new Dependency(Constants.FirstInstallerDependencyName, FirstInstallerDependencyPath),
+                    new Dependency(Constants.SecondInstallerDependencyName, SecondInstallerDependencyPath)
+                } 
+            },
+            {Constants.FirstInstallerDependencyName, new List<Dependency> {new Dependency(Constants.DependencyRepoName, CommonDependencyPath) }},
+            {Constants.SecondInstallerDependencyName, new List<Dependency> {new Dependency(Constants.DependencyRepoName, CommonDependencyPath) }},
+        };
+
+        await CreateRepositoryRecursive(CommonInstallerPath, Constants.InstallerRepoName, dependenciesMap);
+        Directory.CreateDirectory(CommonInstallerPath / Constants.PatchesFolderName / Constants.ProductRepoName);
     }
 
     [OneTimeTearDown]
@@ -109,13 +123,8 @@ public class VmrTestsOneTimeSetUp
     private async Task<string> CreateRepositoryRecursive(
         LocalPath repoPath,
         string repoName,
-        IDictionary<string, List<Dependency>>? dependencies = null,
-        bool createVersionDetails = false)
+        IDictionary<string, List<Dependency>>? dependencies = null)
     {
-        Directory.CreateDirectory(repoPath);
-        Directory.CreateDirectory(repoPath / "eng");
-
-        var dependenciesString = new StringBuilder();
         if (dependencies != null && dependencies.ContainsKey(repoName))
         {
             var repoDependencies = dependencies[repoName];
@@ -123,26 +132,21 @@ public class VmrTestsOneTimeSetUp
             {
                 if (!Directory.Exists(dep.Uri))
                 {
-                    string sha = await CreateRepositoryRecursive(dep.Uri, dep.Name, dependencies);
-                    if (createVersionDetails)
-                    {
-                        dependenciesString.AppendLine(
-                            string.Format(
-                                Constants.DependencyTemplate,
-                                new[] { dep.Name, dep.Uri, sha }));
-                    }
+                    await CreateRepositoryRecursive(dep.Uri, dep.Name, dependencies);
                 }
             }
         }
 
-        File.WriteAllText(repoPath / $"{repoName}-file.txt", $"File in {repoName}");
-        if (createVersionDetails)
+        if (!Directory.Exists(repoPath))
         {
-            var versionDetails = string.Format(Constants.VersionDetailsTemplate, dependenciesString);
-            File.WriteAllText(repoPath / VersionFiles.VersionDetailsXml, versionDetails);
-        }
+            Directory.CreateDirectory(repoPath);
+            Directory.CreateDirectory(repoPath / "eng");
 
-        await _gitOperations.InitialCommit(repoPath);
+            File.WriteAllText(repoPath / $"{repoName}-file.txt", $"File in {repoName}");
+            
+            await _gitOperations.InitialCommit(repoPath);
+        }
+        
         return await _gitOperations.GetRepoLastCommit(repoPath);
     }
 }
